+++ wow_dbc/src/lib.rs
@@ -136,9 +136,9 @@ pub trait Indexable: DbcTable {
 
     /// Gets the primary key, if present. Internally this is just [`std::iter::Iterator::find`] since the
     /// items are not guaranteed to be ordered nor even be present.
-    fn get(&self, key: impl Into<Self::PrimaryKey>) -> Option<&Self::Row>;
+    fn get(&self, key: impl TryInto<Self::PrimaryKey>) -> Option<&Self::Row>;
 
     /// Gets the primary key, if present. Internally this is just [`std::iter::Iterator::find`] since the
     /// items are not guaranteed to be ordered nor even be present.
-    fn get_mut(&mut self, key: impl Into<Self::PrimaryKey>) -> Option<&mut Self::Row>;
+    fn get_mut(&mut self, key: impl TryInto<Self::PrimaryKey>) -> Option<&mut Self::Row>;
 }