+++ wow_dbc/src/tbc_tables/lfg_dungeons.rs
@@ -149,13 +149,13 @@ impl DbcTable for LFGDungeons {
 
 impl Indexable for LFGDungeons {
     type PrimaryKey = LFGDungeonsKey;
-    fn get(&self, key: impl Into<Self::PrimaryKey>) -> Option<&Self::Row> {
-        let key = key.into();
+    fn get(&self, key: impl TryInto<Self::PrimaryKey>) -> Option<&Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter().find(|a| a.id.id == key.id)
     }
 
-    fn get_mut(&mut self, key: impl Into<Self::PrimaryKey>) -> Option<&mut Self::Row> {
-        let key = key.into();
+    fn get_mut(&mut self, key: impl TryInto<Self::PrimaryKey>) -> Option<&mut Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter_mut().find(|a| a.id.id == key.id)
     }
 }