+++ wow_dbc/src/vanilla_tables/char_sections.rs
@@ -163,13 +163,13 @@ impl DbcTable for CharSections {
 
 impl Indexable for CharSections {
     type PrimaryKey = CharSectionsKey;
-    fn get(&self, key: impl Into<Self::PrimaryKey>) -> Option<&Self::Row> {
-        let key = key.into();
+    fn get(&self, key: impl TryInto<Self::PrimaryKey>) -> Option<&Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter().find(|a| a.id.id == key.id)
     }
 
-    fn get_mut(&mut self, key: impl Into<Self::PrimaryKey>) -> Option<&mut Self::Row> {
-        let key = key.into();
+    fn get_mut(&mut self, key: impl TryInto<Self::PrimaryKey>) -> Option<&mut Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter_mut().find(|a| a.id.id == key.id)
     }
 }