+++ wow_dbc/src/vanilla_tables/vocal_ui_sounds.rs
@@ -131,13 +131,13 @@ impl DbcTable for VocalUISounds {
 
 impl Indexable for VocalUISounds {
     type PrimaryKey = VocalUISoundsKey;
-    fn get(&self, key: impl Into<Self::PrimaryKey>) -> Option<&Self::Row> {
-        let key = key.into();
+    fn get(&self, key: impl TryInto<Self::PrimaryKey>) -> Option<&Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter().find(|a| a.id.id == key.id)
     }
 
-    fn get_mut(&mut self, key: impl Into<Self::PrimaryKey>) -> Option<&mut Self::Row> {
-        let key = key.into();
+    fn get_mut(&mut self, key: impl TryInto<Self::PrimaryKey>) -> Option<&mut Self::Row> {
+        let key = key.try_into().ok()?;
         self.rows.iter_mut().find(|a| a.id.id == key.id)
     }
 }