+++ wow_dbc/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* BREAKING: Changed `Indexable::get()` and `Indexable::get_mut()` parameters from `Into<PrimaryKey>` to `TryInto<PrimaryKey>`. This will allow more integer types to be used no matter the underlying integer type.
 * BREAKING: Types used are now from the `wow_world_base` crate.
 * BREAKING: Changed `DbcTable::filename()` to associated const `FILENAME`.
   This makes the name available at compile time, as well as better communicating that it's just a constant value.